Buffer directional pad
Because the USB report does not store the directional pad as a bitfield, we need to buffer it so that if the user sets all of the directional pad inputs separately the output is true to their intent.

With this buffering, the directional pad state is stored in the class and encoded into the USB packet on send().

Note that this is *not* compatible with the v1 API dpad/write.